requirements.txt
langchain
streamlit
python-dotenv
openai
duckduckgo-search
langchain_community
python-weather